Separation of Tandem, FeedForward and SinkTree packages
The feedforward package now only contains the arrival bounding code that
is able to backtrack through any feedforward network. The code for
classical tandem analyses (TFA, SFA, PMOO) was split out and put into a
separate "tandem" package. This change also required to move the
AnalysisConfig into the common (logical) parent package as it brings
both parts of an analysis together. The presets class moved there, too.
The sink-tree specific arrival bound was separated into a new sinktree
package as well. Commented code about making it extend the ff
ArrivalBound was removed.
