+++ src/main/java/de/uni_kl/cs/discodnc/bounds/disco/pwaffine/Backlog.java
@@ -31,17 +31,17 @@ package de.uni_kl.cs.discodnc.bounds.disco.pwaffine;
 
 import java.util.ArrayList;
 
+import de.uni_kl.cs.discodnc.AnalysisConfig;
 import de.uni_kl.cs.discodnc.Calculator;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.curves.Curve;
 import de.uni_kl.cs.discodnc.curves.ServiceCurve;
-import de.uni_kl.cs.discodnc.feedforward.AnalysisConfig;
-import de.uni_kl.cs.discodnc.feedforward.arrivalbounds.SinkTree_AffineCurves;
 import de.uni_kl.cs.discodnc.network.server_graph.Flow;
 import de.uni_kl.cs.discodnc.network.server_graph.Server;
 import de.uni_kl.cs.discodnc.network.server_graph.ServerGraph;
 import de.uni_kl.cs.discodnc.network.server_graph.Turn;
 import de.uni_kl.cs.discodnc.numbers.Num;
+import de.uni_kl.cs.discodnc.sinktree.arrivalbounds.SinkTree_AffineCurves;
 
 public class Backlog {
 	private Backlog() {
