+++ src/main/java/de/uni_kl/cs/discodnc/feedforward/AbstractArrivalBound.java
@@ -28,6 +28,7 @@
 
 package de.uni_kl.cs.discodnc.feedforward;
 
+import de.uni_kl.cs.discodnc.AnalysisConfig;
 import de.uni_kl.cs.discodnc.network.server_graph.ServerGraph;
 
 public abstract class AbstractArrivalBound implements ArrivalBound {
