+++ src/main/java/de/uni_kl/cs/discodnc/feedforward/ArrivalBound.java
@@ -30,6 +30,7 @@ package de.uni_kl.cs.discodnc.feedforward;
 
 import java.util.Set;
 
+import de.uni_kl.cs.discodnc.AnalysisConfig;
 import de.uni_kl.cs.discodnc.curves.ArrivalCurve;
 import de.uni_kl.cs.discodnc.network.server_graph.Flow;
 import de.uni_kl.cs.discodnc.network.server_graph.ServerGraph;
